extract blog component
