+++ apps/website/src/components/index.ts
@@ -2,3 +2,4 @@ export * from './layout';
 export * from './seo';
 export * from './social-media';
 export * from './about';
+export * from './blogs';
